pyproject.toml
[project]
name = "ome-zarr-vole-tools"
version = "0.1.0"
description = "Convert microscopy images to OME-Zarr format and generate VolE viewer URLs"
requires-python = ">=3.10"
dependencies = [
    "bioio>=1.0.0",
    "bioio-tifffile>=1.0.0",
    "bioio-ome-tiff>=1.0.0",
    "bioio-czi>=1.0.0",
    "bioio-lif>=1.0.0",
    "bioio-nd2>=1.0.0",
    "bioio-dv>=1.0.0",
    "ome-zarr>=0.9.0",
    "zarr>=2.16,<3",
    "dask[array]",
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "tifffile>=2023.0",
    "numpy",
]

[project.scripts]
ome-zarr-vole-tools = "ome_zarr_vole_tools.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ome_zarr_vole_tools"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "tifffile>=2023.0",
    "numpy",
]
